requirements.txt
# Core dependencies
streamlit>=1.30.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langgraph>=1.0.0

# HTTP requests
requests>=2.31.0

# Web search
duckduckgo-search>=8.0.0
